Edited dtos

+++ src/main/java/tetrad/virtualmuseum/DTO/ImageDTO.java
@@ -7,7 +7,7 @@ import org.springframework.lang.Nullable;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record ImageDTO(
 
-        @JsonProperty("id") int personalGalleryId,
+        @JsonProperty("id") Long personalGalleryId,     
 
         @JsonProperty("title") String title,
         @JsonProperty("place_of_origin") String placeOfOrigin,
@@ -20,7 +20,7 @@ public record ImageDTO(
 
         @Nullable @JsonProperty("gallery_title") String galleryTitle,
         @JsonProperty("artist_display") String artistDisplay,
-        @Nullable @JsonProperty("gallery_id") Integer galleryId,
+        @Nullable @JsonProperty("gallery_id") Long galleryId,
         @JsonProperty("image_id") String imageId
 ) {
